Support polymorphic associations.
Breaking rewrite to store "from-one" data in a hash. as opposed to
prefixed fields directly on the model. E.g. what was @head.person_name
is now @head.person_fields["name"].

The decision was made to go this way to ensure the implementations of
one's, many's, and polymorphic association stay fairly consistent.