Fix specs that broke in Mongoid 3.
Most of these had to do with the millisecond precision of timestamps
causing time diffs where times should be equal. Removing explicit
created at setting fixes that.

Changed a spec where setting a string into an integer field causes
a 0, not the string, to then exist in the field.

Changed any usage of the @head.below_people = [@person] assignment. This
throws an exception in Mongoid 3, but using the << operator works in 2
and 3.